Deleted dedicated directories of all components, instead moved the components directly to the components directory.
